Move shouldShowProperty function to utils.js
