+++ src/editor/components/modals/ScreenshotModal/ScreenshotProperties.component.jsx
@@ -2,12 +2,12 @@ import { useCallback, useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
 import styles from './ScreenshotModal.module.scss';
 import PropertyRow from '../../elements/PropertyRow.js';
-import { shouldShowProperty } from '../../elements/Component.js';
 import { makeScreenshot } from '@/editor/lib/SceneUtils.js';
 import { debounce } from 'lodash-es';
 import { useAuthContext } from '../../../contexts';
 import useStore from '@/store';
 import Events from '../../../lib/Events.js';
+import { shouldShowProperty } from '../../../lib/utils.js';
 
 const ScreenshotProperties = ({ entity }) => {
   const { currentUser } = useAuthContext();
